fix: resolved performance issues by switching to useSubscribeSuspense + useFind for Dashboard as well as modularising it.
